+++ packages/expo-gl/CHANGELOG.md
@@ -11,6 +11,8 @@
 ### 💡 Others
 
 - Methods of the legacy `EXGLObjectManager` module are now exported from the existing native module in Swift. ([#41171](https://github.com/expo/expo/pull/41171) by [@tsapeta](https://github.com/tsapeta))
+- [android] rename module to align with iOS ([#41193](https://github.com/expo/expo/pull/41193) by [@Ubax](https://github.com/Ubax))
+
 
 ### ⚠️ Notices